scripts/requirements.txt
# Requirements for Hosted Agent registration scripts
# Install with: pip install -r scripts/requirements.txt

azure-ai-projects>=2.0.0b3
azure-identity>=1.17.0
